analytics/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
clickhouse-connect==0.9.1
idna==3.10
lz4==4.4.4
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
urllib3==2.5.0
zstandard==0.25.0
